# Conflicts:
#	server/src/main/java/codestates/frogroup/indiego/domain/show/dto/ShowCommentDto.java
#	server/src/main/java/codestates/frogroup/indiego/domain/show/dto/ShowReservationDto.java

+++ server/src/main/java/codestates/frogroup/indiego/domain/article/entity/dto/ArticleDto.java
@@ -72,7 +72,7 @@ public class ArticleDto {
 
         private List<ArticleCommentDto.Response> articleComments;
 
-        private Long articleCommentCount;
+        private Integer articleCommentCount;
 
     }
